make one directory per one module
