+++ test/simulator/pauli_circuit/test_add_gate.py
@@ -5,12 +5,12 @@ from quantestpy.exceptions import QuantestPyTestCircuitError
 from quantestpy.simulator.pauli_circuit import _IMPLEMENTED_GATES
 
 
-class TestPauliCircuitAddGate(unittest.TestCase):
+class TestAddGate(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.simulator.test_pauli_circuit_add_gate
+    $ python -m unittest test.simulator.pauli_circuit.test_add_gate
     ..
     ----------------------------------------------------------------------
     Ran 2 tests in 0.001s
